+++ ext/graphics/circle.c
@@ -483,17 +483,14 @@ static VALUE Circle_get_global_bounds(VALUE self) {
  * @return [nil]
  */
 static VALUE Circle_draw(VALUE self, VALUE rb_target, VALUE rb_state) {
-    if (!rb_obj_is_kind_of(rb_target, Get_Klass_Target())) {
-        raise_invalid_argument_class(Get_Klass_Target());
-    }
+    TargetView view = Get_RenderTarget_View(rb_target);
 
     if (!rb_obj_is_kind_of(rb_state, Get_Klass_RenderState())) {
         raise_invalid_argument_class(Get_Klass_RenderState());
     }
 
-    TARGET_DRAW(Get_Target_Struct(rb_target), sfRenderWindow_drawCircleShape,
-                sfRenderTexture_drawCircleShape, Get_Circle_Shape(self),
-                Get_RenderState_Struct(rb_state));
+    TARGET_DRAW(view, sfRenderWindow_drawCircleShape, sfRenderTexture_drawCircleShape,
+                Get_Circle_Shape(self), Get_RenderState_Struct(rb_state));
 
     return Qnil;
 }