+++ ext/graphics/target.c
@@ -5,12 +5,14 @@
 
 #include "graphics/render_state.h"
 #include "graphics/render_texture.h"
+#include "graphics/render_window.h"
 #include "graphics/view.h"
 #include "window/window.h"
 #include "core/exceptions.h"
 #include "core/macros.h"
 
 static VALUE rb_cTarget;
+static VALUE rb_mRenderTarget;
 
 static Target* Target_create(TargetType type, void* handle, VALUE rb_source) {
     Target* target = malloc(sizeof(Target));
@@ -131,10 +133,38 @@ static VALUE RenderTarget_get_view(VALUE self) {
     return Get_Casting_View(view);
 }
 
+TargetView Get_RenderTarget_View(VALUE self) {
+    TargetView view;
+
+    if (rb_obj_is_kind_of(self, Get_Klass_Target())) {
+        Target* target = Get_Target_Struct(self);
+
+        view.type = target->type;
+        view.handle = target->handle;
+        return view;
+    }
+
+    if (rb_obj_is_kind_of(self, rb_mRenderTarget)) {
+        if (rb_obj_is_kind_of(self, Get_Klass_RenderWindow())) {
+            view.type = SFML_TARGET_WINDOW;
+            view.handle = Get_Window_Struct(self);
+        } else {
+            view.type = SFML_TARGET_TEXTURE;
+            view.handle = Get_RenderTexture_Struct(self);
+        }
+        return view;
+    }
+
+    rb_raise(rb_eArgError, "expected a RenderTarget (RenderWindow or RenderTexture)");
+}
+
 /* Document-class: SFML::Target
- * A generic handle onto whatever can be drawn to -- a Window or a
+ * A legacy generic handle onto whatever can be drawn to -- a Window or a
  * RenderTexture -- used by Drawable#draw so drawable objects don't need to
  * know which concrete kind of target they're being drawn onto.
+ *
+ * New code should pass a SFML::RenderWindow or SFML::RenderTexture (both
+ * include SFML::RenderTarget) straight to a drawable's #draw.
  *
  * @!attribute view
  *   The target's current view.
@@ -142,6 +172,7 @@ static VALUE RenderTarget_get_view(VALUE self) {
  */
 void Init_Target(VALUE rb_mSFML) {
     rb_cTarget = rb_define_class_under(rb_mSFML, "Target", rb_cObject);
+    rb_mRenderTarget = rb_define_module_under(rb_mSFML, "RenderTarget");
 
     rb_define_alloc_func(rb_cTarget, RenderTarget_alloc);
 
@@ -160,6 +191,10 @@ VALUE Get_Klass_Target(void) {
     return rb_cTarget;
 }
 
+VALUE Get_Module_RenderTarget(void) {
+    return rb_mRenderTarget;
+}
+
 Target* Get_Target_Struct(VALUE self) {
     Target* ptr;
     TypedData_Get_Struct(self, Target, &RenderTarget_data_type, ptr);
@@ -173,13 +208,4 @@ VALUE Get_New_Target(VALUE rb_window) {
 
     return RenderTarget_wrap(Get_Klass_Target(), SFML_TARGET_WINDOW, Get_Window_Struct(rb_window),
                              rb_window);
-}
-
-VALUE Get_New_Target_From_RenderTexture(VALUE rb_render_texture) {
-    if (!rb_obj_is_kind_of(rb_render_texture, Get_Klass_RenderTexture())) {
-        raise_invalid_argument_class(Get_Klass_RenderTexture());
-    }
-
-    return RenderTarget_wrap(rb_cTarget, SFML_TARGET_TEXTURE,
-                             Get_RenderTexture_Struct(rb_render_texture), rb_render_texture);
 }